# internal/strategies: use plain string concatenation in normalizedValue