Simple string concatenation provides the best performance when the
number of substrings is "statically" known (as for this case).

Go 1.13.3. Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz.

```
name                                            old time/op    new time/op    delta
NormalizedValue/value_less_than_32_bytes-12        305ns ± 4%     150ns ± 8%  -50.78%  (p=0.000 n=9+9)
NormalizedValue/value_greater_than_32_bytes-12     351ns ±12%     243ns ±11%  -30.90%  (p=0.000 n=10+10)

name                                            old alloc/op   new alloc/op   delta
NormalizedValue/value_less_than_32_bytes-12         144B ± 0%       96B ± 0%  -33.33%  (p=0.000 n=10+10)
NormalizedValue/value_greater_than_32_bytes-12      208B ± 0%      176B ± 0%  -15.38%  (p=0.000 n=10+10)

name                                            old allocs/op  new allocs/op  delta
NormalizedValue/value_less_than_32_bytes-12         5.00 ± 0%      2.00 ± 0%  -60.00%  (p=0.000 n=10+10)
NormalizedValue/value_greater_than_32_bytes-12      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
```