@tormoder 2019-10-24
Rebased on v3.

Unrelated: I'm seeing variants of this failure both locally and on Travis (non-deterministic).  Known issue?

```
--- FAIL: TestMetrics_SendMetricsFail (0.51s)
    metrics_test.go:210: 
        	Error Trace:	metrics_test.go:202
        	            				metrics_test.go:210
        	Error:      	Not equal: 
        	            	expected: 1
        	            	actual  : 0
        	Test:       	TestMetrics_SendMetricsFail
```