@frekui 2019-10-25
I'm working on a fix for the failure in TestMetrics_SendMetricsFail. Thanks for the report @tormoder!