@jrbarron 2019-10-25
The `TestMetrics_SendMetricsFail` is visible on all branches now so I'll just merge this one and wait for the fix from @frekui.