@frekui 2019-10-25
I created #63 to fix the test failure.